Use .intValue in creating dates.
Former-commit-id: 6683aae089123a1a1ed3be465e47e10dab5e4ce1

+++ jolieJavaServices/src/joliex/db/impl/NamedStatementParser.java
@@ -75,8 +75,8 @@ public class NamedStatementParser
 						statement.setDate( index,
 							Date.valueOf(
 								date.getFirstChild( "year").intValue()
-								+ "-" + date.getFirstChild( "month" )
-								+ "-" + date.getFirstChild( "day" )
+								+ "-" + date.getFirstChild( "month" ).intValue()
+								+ "-" + date.getFirstChild( "day" ).intValue()
 							)
 						);
 					}
